Refactor scene editor utilities and factory functions
