+++ src/main/kotlin/no/njoh/pulseengine/widgets/editor/MenuBarButton.kt
@@ -1,0 +1,5 @@
+package no.njoh.pulseengine.widgets.editor
+
+data class MenuBarButton(val labelText: String, val items: List<MenuBarItem>)
+
+data class MenuBarItem(val labelText: String, val onClick: () -> Unit)
